feat: Refactor sniffer structure
