--- sniffer/include/get_header_len.h
@@ -1,4 +1,0 @@
-#ifndef INCLUDE_GET_HEADER_LEN_H
-#define INCLUDE_GET_HEADER_LEN_H
-
-#endif
