+++ sniffer/include/packet_handler.h
@@ -1,4 +1,8 @@
 #ifndef INCLUDE_PACKET_HANDLER_H
 #define INCLUDE_PACKET_HANDLER_H
 
-#endif
+#include "common.h"
+
+void packet_handler(u_char *user, const struct pcap_pkthdr *packethdr, const u_char *packetptr);
+
+#endif // INCLUDE_PACKET_HANDLER_H
