+++ sniffer/src/globals.c
@@ -1,0 +1,5 @@
+#include "common.h"
+
+extern pcap_t* handle = NULL;
+extern int linkhdrlen = 0;
+extern int packets = 0;
